fix: move user preferences utils to dedicated file
