+++ test/unit/specs/util/myKivaUtils.spec.js
@@ -1,11 +1,7 @@
 import {
 	hasLoanFunFactFootnote,
-	createUserPreferences,
-	updateUserPreferences,
 	getIsMyKivaEnabled,
 	fetchPostCheckoutAchievements,
-	createUserPreferencesMutation,
-	updateUserPreferencesMutation,
 	setGuestAssignmentCookie,
 	checkGuestAssignmentCookie,
 	GUEST_ASSIGNMENT_COOKIE,
@@ -16,35 +12,6 @@ import logReadQueryError from '#src/util/logReadQueryError';
 import * as experimentUtils from '#src/util/experiment/experimentUtils';
 
 vi.mock('#src/util/logReadQueryError');
-
-const userPreferenceDataMock = {
-	my: {
-		userPreference: {
-			id: 1,
-			preferences: ''
-		}
-	}
-};
-
-const createPreferenceDataMock = {
-	my: {
-		createUserPreferences: {
-			id: 1,
-			preferences: ''
-		}
-	}
-};
-
-const updatedPreferenceDataMock = {
-	my: {
-		userPreference: {
-			id: 1,
-			preferences: {
-				test: 'test'
-			}
-		}
-	}
-};
 
 describe('myKivaUtils.js', () => {
 	describe('hasLoanFunFactFootnote', () => {
@@ -162,55 +129,6 @@ describe('myKivaUtils.js', () => {
 			await fetchPostCheckoutAchievements(apolloMock, loanIds);
 
 			expect(logReadQueryError).toHaveBeenCalledWith(error, 'myKivaUtils postCheckoutAchievementsQuery');
-		});
-	});
-
-	describe('createUserPreferences', () => {
-		it('should create user preferences as expected', async () => {
-			const apolloMock = {
-				mutate: vi.fn().mockReturnValueOnce(Promise.resolve({ data: createPreferenceDataMock }))
-			};
-			const preferences = { test: 'test' };
-
-			const newUserPreference = await createUserPreferences(apolloMock, preferences);
-
-			expect(newUserPreference).toEqual({
-				data: createPreferenceDataMock
-			});
-			expect(apolloMock.mutate).toHaveBeenCalledTimes(1);
-			expect(apolloMock.mutate).toHaveBeenCalledWith({
-				mutation: createUserPreferencesMutation,
-				variables: { preferences: JSON.stringify(preferences) },
-			});
-		});
-	});
-
-	describe('updateUserPreferences', () => {
-		it('should update user preferences as expected', async () => {
-			const apolloMock = {
-				mutate: vi.fn()
-					.mockReturnValueOnce(Promise.resolve({ data: updatedPreferenceDataMock }))
-			};
-
-			const newUserPreference = await updateUserPreferences(
-				apolloMock,
-				userPreferenceDataMock.my.userPreference,
-				// Pass existing preferences to ensure they are merged correctly
-				{ test: 'test' },
-				{ new: 'new' },
-			);
-
-			expect(newUserPreference).toEqual({
-				data: updatedPreferenceDataMock
-			});
-			expect(apolloMock.mutate).toHaveBeenCalledTimes(1);
-			expect(apolloMock.mutate).toHaveBeenCalledWith({
-				mutation: updateUserPreferencesMutation,
-				variables: {
-					updateUserPreferencesId: userPreferenceDataMock.my.userPreference.id,
-					preferences: JSON.stringify({ test: 'test', new: 'new' }),
-				},
-			});
 		});
 	});
